Added support for Redox OS and refactoring.
This commit adds support for Redox OS. Along with that it does a bit of
refactoring so UNIX code and Redox code live separately and the common
stuff is defined only once. Rationale being that in the first
implementation (w/o separate sub-modules I ended up with many, many
conditional compilation annotations and lots of duplicate functions and
methods definitions). I tried to preserved the original structure of the
library for UNIX and preserve the same layout of the entities and the
API son thing gets broken by this. This is part of the ongoing effort to
support `exa`.

Details

- Separated UNIX and Redox implementations in a new `base` module
(previously in `base.rs`).
- Separated UNIX and Redox implementations in a new `switch` module
(previously in `switch.rs`).
- Added `redox_users` and `redox_syscall` dependencies (only on Redox
targets).
- Added public(crate) visibility to fields in `User` and `Group`.
- Implemented OS extensions for Redox.
- Implemented `From` trait from `redox_users::User` to `User`.
- Fixed imports due to changes in mod hierarchy.