Fix doctest warnings
A recent version of the compiler started warning about `norun` in the cache doctests, saying it should be `no_run`. These doctests are not really documentation tests — they're more development notes with blocks of Rust code in them. As such, changing them to `no_run` makes them fail.

The first doctest was salvageable, as it only contained a syntax error and a reference to a function that wasn't updated when it was renamed. The other one had to be changed to `text`, as it wasn't intended to pass at all.