Use Arc<OsStr> to avoid double redirection
This commit changes all instances of Arc<OsString> to Arc<OsStr>, which accomplishes much the same thing, without having to redirect twice. This includes some public interfaces, making this a breaking change.

The original code was written so long ago that I wasn't fully accustomed to dynamically-sized types. Recently, Clippy added a warning (clippy::rc_buffer) that pointed it out, so I changed them. This also removes a couple of unwraps, as we can get a reference value out of Arcs instead of having to use Arc::try_unwrap.